move process_data functions to new folder and write logging file instead of print
