+++ run.py
@@ -1,7 +1,6 @@
 # from crawler.winmart.demo import main
-# from crawler.bhx.demo import run_sync
-from crawler.winmart.demo import run_sync
+from crawler.bhx.demo import run_sync
+# from crawler.winmart.demo import run_sync
 
 if __name__ == "__main__":
-    # asyncio.run(WinMartFetcher().run())
     run_sync()
